Rename IndicatorsUtils to IndicatorsTypes.
